requirements.txt
deterrers-api>=0.3
PyYAML
Click
